Additions to modular arithmetic test cases.
This is part of the effort to add subjective limits to the
mod_exp host function (see eosnetworkfoundation/mandel#361).

Adds a few additional test vectors to existing modexp test case.

Adds a modexp_benchmarking test case which is not actually a test
but rather a way to get measurements on the time it takes to do
modular exponentiation with varying bit sizes to figure out what the
appropriate limit should be for the mod_exp host function.